requirements.txt
pandas==2.2.3
numpy==2.2.4
scikit-learn>=1.5.0
xgboost>=2.1.0
joblib>=1.4.0
matplotlib>=3.9.0
PyYAML>=6.0
